fix overflow due to wrong return type

+++ storage/preferences/src/main/java/de/danoeh/antennapod/storage/preferences/UserPreferences.java
@@ -458,7 +458,7 @@ public abstract class UserPreferences {
     }
 
     public static long getUpdateInterval() {
-        return Integer.parseInt(prefs.getString(PREF_UPDATE_INTERVAL_MINUTES, "720"));
+        return Long.parseLong(prefs.getString(PREF_UPDATE_INTERVAL_MINUTES, "720"));
     }
 
     public static void setUpdateInterval(long interval) {
